+++ bin/build-storybook/copy-package-dist.mjs
@@ -1,4 +1,3 @@
-import { existsSync } from 'node:fs';
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import { argv, cwd } from 'node:process';
@@ -17,16 +16,9 @@ async function main(args) {
         })
         .parse();
 
-    const target = path.join(to, path.basename(cwd()));
+    const basename = path.basename(cwd());
 
-    await fs.cp('dist', target, { recursive: true });
-
-    /**
-     * Need to resolve dependencies correctly, so copy `node_modules`
-     */
-    if (existsSync('node_modules')) {
-        await fs.cp('node_modules', path.join(target, 'node_modules'), { recursive: true });
-    }
+    await fs.cp('dist', path.join(to, basename), { recursive: true });
 }
 
 await main(hideBin(argv));