+++ packages/bottom-sheet/src/component.tsx
@@ -66,6 +66,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             containerClassName,
             containerProps,
             headerClassName,
+            headerContentClassName,
             footerClassName,
             addonClassName,
             closerClassName,
@@ -123,6 +124,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             iOSLock = false,
             virtualKeyboard = false,
             colors = 'default',
+            preventScrollOnSwipe,
         },
         ref,
     ) => {
@@ -186,6 +188,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             headerOffset,
             colors,
             className: cn(headerClassName, colorStyle.hasContent),
+            contentClassName: headerContentClassName,
             addonClassName,
             closerClassName,
             backButtonClassName: backerClassName,
@@ -494,6 +497,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             trackMouse: swipeable,
             trackTouch: swipeable,
             delta: swipeThreshold,
+            preventScrollOnSwipe,
         });
 
         const handleExited = (node: HTMLElement) => {