+++ packages/bottom-sheet/src/types.ts
@@ -1,4 +1,5 @@
 import { type HTMLAttributes, type ReactElement, type ReactNode, type RefObject } from 'react';
+import { type SwipeableProps } from 'react-swipeable';
 import { type HandledEvents } from 'react-swipeable/es/types';
 import { type TransitionProps } from 'react-transition-group/Transition';
 
@@ -86,6 +87,11 @@ export type BottomSheetProps = {
      * Дополнительный класс шапки
      */
     headerClassName?: string;
+
+    /**
+     * Дополнительный класс для контента шапки
+     */
+    headerContentClassName?: string;
 
     /**
      * Дополнительный класс футера
@@ -369,4 +375,10 @@ export type BottomSheetProps = {
      * @default default
      */
     colors?: 'default' | 'inverted';
+
+    /**
+     * Предотвращать скролл во время свайпа
+     * @default false
+     */
+    preventScrollOnSwipe?: SwipeableProps['preventScrollOnSwipe'];
 };