+++ packages/plate/src/components/base-plate/Component.tsx
@@ -166,11 +166,6 @@ export type BasePlateProps = {
      * Основные стили компонента.
      */
     styles?: { [key: string]: string };
-
-    /**
-     * Кастомный радиус скругления (px, rem, em и т.д.)
-     */
-    borderRadius?: string;
 };
 
 /* eslint-disable complexity */
@@ -205,7 +200,6 @@ export const BasePlate = forwardRef<HTMLDivElement, BasePlateProps>(
             styles = {},
             background,
             borderColor,
-            borderRadius,
         },
         ref,
     ) => {
@@ -326,7 +320,6 @@ export const BasePlate = forwardRef<HTMLDivElement, BasePlateProps>(
                 style={{
                     ...setCustomViewColors(),
                     ...setCustomViewRectangleBorderColor(),
-                    ...(borderRadius && { borderRadius }),
                 }}
                 onClick={handleClick}
                 onKeyDown={handleClick}