+++ packages/stepped-progress-bar/src/Component.tsx
@@ -24,7 +24,10 @@ export type SteppedProgressBarView =
     | 'primary'
     | 'accent';
 
-export type SteppedProgressBarProps = {
+export type CustomProgressBarView = { background: string };
+export type SteppedProgressBarViewValue = SteppedProgressBarView | CustomProgressBarView;
+
+export interface SteppedProgressBarProps {
     /**
      * Общее количество шагов
      */
@@ -43,7 +46,7 @@ export type SteppedProgressBarProps = {
     /**
      * Цвет заполнения
      */
-    view?: SteppedProgressBarView | SteppedProgressBarView[];
+    view?: SteppedProgressBarViewValue | SteppedProgressBarViewValue[];
 
     /**
      * Идентификатор для систем автоматизированного тестирования
@@ -59,7 +62,7 @@ export type SteppedProgressBarProps = {
      * Дополнительный класс
      */
     className?: string;
-};
+}
 
 export const SteppedProgressBar: FC<SteppedProgressBarProps> = ({
     maxStep,
@@ -70,25 +73,26 @@ export const SteppedProgressBar: FC<SteppedProgressBarProps> = ({
     colors = 'default',
     className,
 }) => {
-    const validMaxSteps = maxStep <= 0 ? 1 : maxStep;
-    const isViewString = typeof view === 'string';
+    const validMaxSteps = Math.max(1, maxStep);
+    const isViewArray = Array.isArray(view);
+    const currentColors = colorStyles[colors];
 
     return (
         <div className={cn(styles.component, className)} data-test-id={dataTestId}>
             <div className={styles.stepsContainer}>
-                {Array.from(Array(validMaxSteps), (_, index) => (
+                {Array.from({ length: validMaxSteps }, (_, index) => (
                     <StepBar
                         key={index}
                         isDone={index < step}
-                        view={(isViewString ? view : view?.[index]) as SteppedProgressBarView}
-                        classNameStep={colorStyles[colors].bar}
+                        view={isViewArray ? view[index] : view}
+                        classNameStep={currentColors.bar}
                     />
                 ))}
             </div>
             {description && (
                 <Text
                     tag='div'
-                    className={cn(styles.description, colorStyles[colors].description)}
+                    className={cn(styles.description, currentColors.description)}
                     view='primary-small'
                 >
                     Шаг {step} из {maxStep}: {description}