+++ packages/calendar-range/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @alfalab/core-components-calendar-range
 
+## 8.0.7
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.6
+    - @alfalab/core-components-calendar-input@11.0.7
+
+## 8.0.6
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.5
+    - @alfalab/core-components-calendar-input@11.0.6
+
+## 8.0.5
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-shared@1.1.1
+    - @alfalab/core-components-calendar@8.0.4
+    - @alfalab/core-components-calendar-input@11.0.5
+    - @alfalab/core-components-date-input@5.0.3
+
+## 8.0.4
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-calendar@8.0.3
+    - @alfalab/core-components-calendar-input@11.0.4
+
 ## 8.0.3
 
 ### Patch Changes