+++ packages/comment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @alfalab/core-components-comment
 
+## 3.0.3
+
+### Patch Changes
+
+- Обновлены зависимости
+    - @alfalab/core-components-typography@5.0.3
+
 ## 3.0.2
 
 ### Patch Changes